requirements.txt
streamlit>=1.40.0
yfinance>=0.2.40
finnhub-python>=2.4.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
plotly>=5.0.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0
PyPDF2>=3.0.0
youtube-transcript-api>=0.6.0
gnews>=0.3.0
pandas-datareader>=0.10.0
pandas-ta>=0.4.0
requests-cache>=1.0.0
supabase>=2.0.0
ruff
pytest
